+++ skymapper/map.py
@@ -803,74 +803,3 @@ class Map():
         self.ax.set_xlim(xlim_)
         self.ax.set_ylim(ylim_)
         return artist
-
-
-##### Start of free methods #####
-
-def getOptimalConicProjection(ra, dec, proj_class=None, ra0=None, dec0=None):
-    """Determine optimal configuration of conic map.
-
-    As a simple recommendation, the standard parallels are chosen to be 1/7th
-    closer to dec0 than the minimum and maximum declination in the data
-    (Snyder 1987, page 99).
-
-    If proj_class is None, it will use AlbersEqualAreaProjection.
-
-    Args:
-        ra: list of rectascensions
-        dec: list of declinations
-        proj_class: constructor of projection class
-        ra0: if not None, use this as reference RA
-        dec0: if not None, use this as reference Dec
-
-    Returns:
-        proj_class that best holds ra/dec
-    """
-
-    if ra0 is None:
-        ra_ = np.array(ra)
-        ra_[ra_ > 180] -= 360
-        ra_[ra_ < -180] += 360
-        # weight more towards the poles because that decreases distortions
-        ra0 = (ra_ * dec).sum() / dec.sum()
-
-    if dec0 is None:
-        dec0 = np.median(dec)
-
-    # determine standard parallels for AEA
-    dec1, dec2 = dec.min(), dec.max()
-    # move standard parallels 1/6 further in from the extremes
-    # to minimize scale variations (Snyder 1987, section 14)
-    delta_dec = (dec0 - dec1, dec2 - dec0)
-    dec1 += delta_dec[0]/7
-    dec2 -= delta_dec[1]/7
-
-    if proj_class is None:
-        proj_class = AlbersEqualAreaProjection
-    return proj_class(ra0, dec0, dec1, dec2)
-
-
-def makeMapNice(fig, ax, proj, dec, sep=5, bgcolor="#aaaaaa", cb_collection=None, cb_label=""):
-    # add lines and labels for meridians/parallels
-    meridians = np.arange(-90, 90+sep, sep)
-    parallels = np.arange(0, 360+sep, sep)
-    setMeridianPatches(ax, proj, meridians, linestyle='-', lw=0.5, alpha=0.2, zorder=2)
-    setParallelPatches(ax, proj, parallels, linestyle='-', lw=0.5, alpha=0.2, zorder=2)
-    setMeridianLabels(ax, proj, meridians, loc="left", fmt=pmDegFormatter)
-    if dec.mean() > 0:
-        setParallelLabels(ax, proj, parallels, loc="bottom")
-    else:
-        setParallelLabels(ax, proj, parallels, loc="top")
-
-    if bgcolor is not None:
-        ax.set_facecolor(bgcolor)
-
-    # add colorbar
-    if cb_collection is not None:
-        from mpl_toolkits.axes_grid1 import make_axes_locatable
-        divider = make_axes_locatable(ax)
-        cax = divider.append_axes("right", size="2%", pad=0.0)
-        cb = fig.colorbar(cb_collection, cax=cax)
-        cb.set_label(cb_label)
-        cb.solids.set_edgecolor("face")
-    fig.tight_layout()
